// supabase/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      calls: {
        Row: {
          created_at: string
          endpoint_id: number
          error_message: string | null
          id: number
          is_error: boolean
          method: Database["public"]["Enums"]["http_method_type"]
          response: Json | null
          response_code: number
          slug: string
        }
        Insert: {
          created_at?: string
          endpoint_id: number
          error_message?: string | null
          id?: number
          is_error?: boolean
          method: Database["public"]["Enums"]["http_method_type"]
          response?: Json | null
          response_code?: number
          slug: string
        }
        Update: {
          created_at?: string
          endpoint_id?: number
          error_message?: string | null
          id?: number
          is_error?: boolean
          method?: Database["public"]["Enums"]["http_method_type"]
          response?: Json | null
          response_code?: number
          slug?: string
        }
        Relationships: [
          {
            foreignKeyName: "calls_endpoint_id_fkey"
            columns: ["endpoint_id"]
            isOneToOne: false
            referencedRelation: "endpoints"
            referencedColumns: ["id"]
          },
        ]
      }
      endpoints: {
        Row: {
          created_at: string
          description: string | null
          id: number
          name: string
          slug: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: number
          name: string
          slug?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string
          slug?: string | null
        }
        Relationships: []
      }
      query_strings: {
        Row: {
          call_id: number
          created_at: string
          id: number
          key: string
        }
        Insert: {
          call_id: number
          created_at?: string
          id?: number
          key: string
        }
        Update: {
          call_id?: number
          created_at?: string
          id?: number
          key?: string
        }
        Relationships: [
          {
            foreignKeyName: "query_strings_call_id_fkey"
            columns: ["call_id"]
            isOneToOne: false
            referencedRelation: "calls"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      http_method_type: "GET" | "POST" | "PUT" | "PATCH" | "DELETE"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never